Refactor: 메인페이지 보완2
- 메인페이지 보완2
   - MainService @Slf4j 어노테이션 추가
   - 작성자가 없을 경우 null 처리 추가

+++ src/main/java/com/hotsix/iAmNotAlone/domain/main/model/dto/PostProjectionMainDto.java
@@ -16,7 +16,7 @@ public class PostProjectionMainDto {
     private String createdAt;
     private Long memberId;
     private String nickName;
-    private int gender;
+    private Long gender;
     private String userFile;
     private int commentCount;
     private boolean likesFlag;
